perf: split observatory shell bundles
