+++ packages/phlo-observatory/src/phlo_observatory/src/v2/shell/V2CommandPalette.tsx
@@ -1,0 +1,525 @@
+import { useNavigate } from '@tanstack/react-router'
+import { Command as CommandPrimitive } from 'cmdk'
+import {
+  Activity,
+  Boxes,
+  CirclePlay,
+  Clipboard,
+  Database,
+  GitBranch,
+  LayoutDashboard,
+  ListChecks,
+  Logs,
+  Search,
+  Server,
+  Settings,
+} from 'lucide-react'
+import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
+import type { ReactNode } from 'react'
+
+import type {
+  V2CapabilityPage,
+  V2ResourceResult,
+  V2SearchResult,
+  V2Service,
+  V2Table,
+} from '@/v2/api/types'
+import { getV2Services, getV2TableRecords, searchV2 } from '@/v2/api/resources'
+import { loadCachedResource } from '@/v2/routes/liveResource'
+
+const commandGroupLimit = 6
+
+const iconByPageId: Record<string, typeof LayoutDashboard> = {
+  overview: LayoutDashboard,
+  services: Server,
+  operations: Activity,
+  runs: CirclePlay,
+  data: Database,
+  assets: Boxes,
+  workflows: Clipboard,
+  issues: ListChecks,
+  quality: ListChecks,
+  logs: Logs,
+  branches: GitBranch,
+  extensions: Settings,
+  storage: Database,
+  observability: Activity,
+  governance: Settings,
+  catalog: Boxes,
+  apis: Server,
+  bi: LayoutDashboard,
+  settings: Settings,
+}
+
+export function V2CommandPalette({
+  navItems,
+  onClose,
+}: {
+  navItems: Array<V2CapabilityPage>
+  onClose: () => void
+}) {
+  const navigate = useNavigate()
+  const searchInputRef = useRef<HTMLInputElement | null>(null)
+  const [query, setQuery] = useState('')
+  const [results, setResults] = useState<
+    V2ResourceResult<Array<V2SearchResult>>
+  >({
+    data: null,
+    error: null,
+  })
+  const [commandTables, setCommandTables] = useState<
+    V2ResourceResult<Array<V2Table>>
+  >({
+    data: null,
+    error: null,
+  })
+  const [commandServices, setCommandServices] = useState<
+    V2ResourceResult<Array<V2Service>>
+  >({
+    data: null,
+    error: null,
+  })
+
+  useEffect(() => {
+    const timer = window.setTimeout(() => searchInputRef.current?.focus(), 0)
+    return () => window.clearTimeout(timer)
+  }, [])
+
+  useEffect(() => {
+    if (query.trim().length < 2) {
+      setResults({ data: null, error: null })
+      return
+    }
+    let cancelled = false
+    const timer = window.setTimeout(() => {
+      void searchV2({ data: { query } }).then((next) => {
+        if (!cancelled) setResults(next)
+      })
+    }, 180)
+    return () => {
+      cancelled = true
+      window.clearTimeout(timer)
+    }
+  }, [query])
+
+  useEffect(() => {
+    let cancelled = false
+    void Promise.all([
+      loadCachedResource('v2:tables', getV2TableRecords, {
+        staleMs: 120_000,
+      }),
+      loadCachedResource('v2:services', getV2Services, {
+        staleMs: 120_000,
+      }),
+    ]).then(([tables, services]) => {
+      if (cancelled) return
+      setCommandTables(tables)
+      setCommandServices(services)
+    })
+    return () => {
+      cancelled = true
+    }
+  }, [])
+
+  const closeSearch = useCallback(() => {
+    setQuery('')
+    onClose()
+  }, [onClose])
+
+  const handleCommandSelect = useCallback(
+    (value: string) => {
+      if (value.startsWith('copy-sql:')) {
+        const tableId = value.replace('copy-sql:', '')
+        const sql = `select * from ${tableId} limit 100`
+        void navigator.clipboard?.writeText(sql)
+        closeSearch()
+        return
+      }
+
+      const href = value.replace('open:', '')
+      closeSearch()
+      void navigate({ to: href })
+    },
+    [closeSearch, navigate],
+  )
+
+  const groupedResults = useMemo(() => {
+    const groups = new Map<string, Array<V2SearchResult>>()
+    const excludedKinds = new Set(['service', 'table', 'dataset'])
+    for (const result of results.data ?? []) {
+      if (excludedKinds.has(result.kind)) continue
+      const key = result.kind || 'result'
+      const group = groups.get(key)
+      if (group) {
+        group.push(result)
+      } else {
+        groups.set(key, [result])
+      }
+    }
+    return Array.from(groups.entries())
+  }, [results.data])
+
+  const tableResults = useMemo(
+    () =>
+      (results.data ?? []).filter(
+        (result) => result.kind === 'table' || result.kind === 'dataset',
+      ),
+    [results.data],
+  )
+
+  const tableMatches = useMemo(() => {
+    const needle = query.trim().toLowerCase()
+    const tables = commandTables.data ?? []
+    if (!needle) return tables.slice(0, 12)
+
+    return tables
+      .filter((table) =>
+        [
+          table.id,
+          table.name,
+          tableLabel(table),
+          table.namespace,
+          table.schema_name,
+          table.branch,
+          table.format,
+          table.asset_id,
+          table.metadata.table,
+          table.metadata.table_name,
+          table.metadata.schema,
+          table.metadata.database,
+          table.metadata.catalog,
+          table.metadata.relation,
+          table.metadata.materialized,
+        ]
+          .filter(Boolean)
+          .some((value) => String(value).toLowerCase().includes(needle)),
+      )
+      .slice(0, 16)
+  }, [commandTables.data, query])
+
+  const serviceMatches = useMemo(() => {
+    const needle = query.trim().toLowerCase()
+    const services = commandServices.data ?? []
+    if (!needle) return services.slice(0, 12)
+
+    return services
+      .filter((service) =>
+        [
+          service.id,
+          service.name,
+          service.kind,
+          service.status,
+          service.profile,
+          service.backend,
+          service.health.state,
+          service.health.message,
+          service.metadata.description,
+        ]
+          .filter(Boolean)
+          .some((value) => String(value).toLowerCase().includes(needle)),
+      )
+      .slice(0, 16)
+  }, [commandServices.data, query])
+
+  const pageMatches = useMemo(() => {
+    const needle = query.trim().toLowerCase()
+    if (!needle) return navItems
+    return navItems.filter((item) =>
+      [item.id, item.label, item.path, ...item.providers]
+        .filter(Boolean)
+        .some((value) => String(value).toLowerCase().includes(needle)),
+    )
+  }, [navItems, query])
+
+  const hasLocalMatches =
+    pageMatches.length > 0 ||
+    tableMatches.length > 0 ||
+    serviceMatches.length > 0
+  const sqlTemplateTargets = tableMatches.length
+    ? tableMatches
+    : tableResults.slice(0, 6).flatMap((result) => {
+        const table = tableFromSearchResult(result)
+        return table ? [table] : []
+      })
+
+  return (
+    <div
+      aria-label="Command search"
+      aria-modal="true"
+      className="phlo-v2-command-overlay"
+      role="dialog"
+    >
+      <button
+        aria-label="Close search"
+        className="phlo-v2-command-backdrop"
+        onClick={onClose}
+        type="button"
+      />
+      <div className="phlo-v2-search-popover">
+        <CommandPrimitive
+          className="phlo-v2-command phlo-v2-command-palette"
+          loop
+          shouldFilter={false}
+        >
+          <div className="phlo-v2-search-field">
+            <Search className="size-4" />
+            <CommandPrimitive.Input
+              aria-label="Search Observatory"
+              id="phlo-v2-command-input"
+              onKeyDown={(event) => {
+                if (event.key === 'Escape') {
+                  event.preventDefault()
+                  onClose()
+                }
+              }}
+              onValueChange={setQuery}
+              placeholder="Search services, assets, tables, checks"
+              ref={searchInputRef}
+              value={query}
+            />
+            <kbd>Esc</kbd>
+          </div>
+          <CommandPrimitive.List className="phlo-v2-command-list">
+            {query.trim().length >= 2 && pageMatches.length > 0 && (
+              <CommandPrimitive.Group
+                className="phlo-v2-command-group"
+                heading={`Pages (${pageMatches.length})`}
+              >
+                {pageMatches.slice(0, commandGroupLimit).map((item) => {
+                  const Icon = iconByPageId[item.id] ?? LayoutDashboard
+                  return (
+                    <CommandPrimitive.Item
+                      className="phlo-v2-command-item"
+                      key={`nav:${item.id}`}
+                      onSelect={handleCommandSelect}
+                      value={`open:${item.path}`}
+                    >
+                      <Icon className="size-4" />
+                      <span>{item.label}</span>
+                      <small>{item.providers.join(', ') || 'core'}</small>
+                    </CommandPrimitive.Item>
+                  )
+                })}
+              </CommandPrimitive.Group>
+            )}
+
+            {query.trim().length < 2 && (
+              <CommandPrimitive.Group
+                className="phlo-v2-command-group"
+                heading="Fast actions"
+              >
+                <CommandPrimitive.Item
+                  className="phlo-v2-command-item"
+                  onSelect={handleCommandSelect}
+                  value="open:/data"
+                >
+                  <Database className="size-4" />
+                  <span>Browse tables</span>
+                  <small>Data</small>
+                </CommandPrimitive.Item>
+                <CommandPrimitive.Item
+                  className="phlo-v2-command-item"
+                  onSelect={handleCommandSelect}
+                  value="open:/assets"
+                >
+                  <Boxes className="size-4" />
+                  <span>Inspect assets and lineage</span>
+                  <small>Assets</small>
+                </CommandPrimitive.Item>
+                <CommandPrimitive.Item
+                  className="phlo-v2-command-item"
+                  onSelect={handleCommandSelect}
+                  value="open:/runs"
+                >
+                  <CirclePlay className="size-4" />
+                  <span>Review orchestration runs</span>
+                  <small>Runs</small>
+                </CommandPrimitive.Item>
+              </CommandPrimitive.Group>
+            )}
+
+            {query.trim().length >= 2 && tableMatches.length > 0 && (
+              <CommandPrimitive.Group
+                className="phlo-v2-command-group"
+                heading={`Tables (${tableMatches.length})`}
+              >
+                {tableMatches.slice(0, commandGroupLimit).map((table) => (
+                  <CommandPrimitive.Item
+                    className="phlo-v2-command-item"
+                    key={`table:${table.id}`}
+                    onSelect={handleCommandSelect}
+                    value={`open:/table/${encodeURIComponent(table.id)}`}
+                  >
+                    <Database className="size-4" />
+                    <span>{tableLabel(table)}</span>
+                    <small>{table.branch ?? table.format ?? 'table'}</small>
+                  </CommandPrimitive.Item>
+                ))}
+                {tableMatches.length > commandGroupLimit && (
+                  <CommandPrimitive.Item
+                    className="phlo-v2-command-item"
+                    onSelect={handleCommandSelect}
+                    value="open:/data"
+                  >
+                    <Database className="size-4" />
+                    <span>Open table browser</span>
+                    <small>{tableMatches.length} matches</small>
+                  </CommandPrimitive.Item>
+                )}
+              </CommandPrimitive.Group>
+            )}
+
+            {query.trim().length >= 2 && serviceMatches.length > 0 && (
+              <CommandPrimitive.Group
+                className="phlo-v2-command-group"
+                heading={`Services (${serviceMatches.length})`}
+              >
+                {serviceMatches.slice(0, commandGroupLimit).map((service) => (
+                  <CommandPrimitive.Item
+                    className="phlo-v2-command-item"
+                    key={`service:${service.id}`}
+                    onSelect={handleCommandSelect}
+                    value="open:/services"
+                  >
+                    <Server className="size-4" />
+                    <span>{service.name}</span>
+                    <small>
+                      {[service.kind, service.status]
+                        .filter(Boolean)
+                        .join(' · ')}
+                    </small>
+                  </CommandPrimitive.Item>
+                ))}
+                {serviceMatches.length > commandGroupLimit && (
+                  <CommandPrimitive.Item
+                    className="phlo-v2-command-item"
+                    onSelect={handleCommandSelect}
+                    value="open:/services"
+                  >
+                    <Server className="size-4" />
+                    <span>Open service directory</span>
+                    <small>{serviceMatches.length} matches</small>
+                  </CommandPrimitive.Item>
+                )}
+              </CommandPrimitive.Group>
+            )}
+
+            {groupedResults.map(([kind, items]) => (
+              <CommandPrimitive.Group
+                className="phlo-v2-command-group"
+                heading={`${commandHeading(kind)} (${items.length})`}
+                key={kind}
+              >
+                {items.slice(0, commandGroupLimit).map((result) => (
+                  <CommandPrimitive.Item
+                    className="phlo-v2-command-item"
+                    key={result.id}
+                    onSelect={handleCommandSelect}
+                    value={`open:${cleanPath(result.href ?? '/')}`}
+                  >
+                    {iconForSearchKind(result.kind)}
+                    <span>{result.label}</span>
+                    <small>{result.summary ?? result.kind}</small>
+                  </CommandPrimitive.Item>
+                ))}
+              </CommandPrimitive.Group>
+            ))}
+
+            {query.trim().length >= 2 && sqlTemplateTargets.length > 0 && (
+              <CommandPrimitive.Group
+                className="phlo-v2-command-group"
+                heading="SQL templates"
+              >
+                {sqlTemplateTargets.slice(0, 4).map((table) => (
+                  <CommandPrimitive.Item
+                    className="phlo-v2-command-item"
+                    key={`sql:${table.id}`}
+                    onSelect={handleCommandSelect}
+                    value={`copy-sql:${table.id}`}
+                  >
+                    <Clipboard className="size-4" />
+                    <span>Copy SELECT from {tableLabel(table)}</span>
+                    <small>Read-only template</small>
+                  </CommandPrimitive.Item>
+                ))}
+              </CommandPrimitive.Group>
+            )}
+
+            {query.trim().length >= 2 &&
+              results.data?.length === 0 &&
+              !hasLocalMatches && (
+                <CommandPrimitive.Empty className="phlo-v2-command-empty">
+                  No results found.
+                </CommandPrimitive.Empty>
+              )}
+            {results.error && (
+              <div className="phlo-v2-command-empty">{results.error}</div>
+            )}
+          </CommandPrimitive.List>
+          <div className="phlo-v2-command-footer">
+            <span>↑↓ navigate</span>
+            <span>↵ select</span>
+            <span>esc close</span>
+          </div>
+        </CommandPrimitive>
+      </div>
+    </div>
+  )
+}
+
+function commandHeading(kind: string): string {
+  const labels: Record<string, string> = {
+    asset: 'Assets',
+    table: 'Tables',
+    service: 'Services',
+    log: 'Logs',
+    operation: 'Operations',
+    branch: 'Branches',
+    extension: 'Extensions',
+    run: 'Runs',
+    quality: 'Quality',
+    setting: 'Settings',
+  }
+  return labels[kind] ?? kind
+}
+
+function iconForSearchKind(kind: string): ReactNode {
+  if (kind === 'asset') return <Boxes className="size-4" />
+  if (kind === 'table') return <Database className="size-4" />
+  if (kind === 'service') return <Server className="size-4" />
+  if (kind === 'log') return <Logs className="size-4" />
+  if (kind === 'branch') return <GitBranch className="size-4" />
+  if (kind === 'run') return <CirclePlay className="size-4" />
+  return <Search className="size-4" />
+}
+
+function tableLabel(table: V2Table): string {
+  const namespace = table.namespace ?? table.schema_name
+  if (!namespace) return table.name
+  return `${namespace}.${table.name}`
+}
+
+function tableFromSearchResult(result: V2SearchResult): V2Table | null {
+  if (!['table', 'dataset'].includes(result.kind)) return null
+  const id = result.id.replace(/^table:/, '').replace(/^dataset:/, '')
+  const labelParts = result.label.split('.')
+  const name = labelParts.at(-1) ?? id
+  const namespace =
+    labelParts.length > 1 ? labelParts.slice(0, -1).join('.') : undefined
+  return {
+    id,
+    name,
+    namespace,
+    schema_name: namespace,
+    branch: undefined,
+    format: undefined,
+    asset_id: undefined,
+    metadata: {},
+  }
+}
+
+function cleanPath(path: string): string {
+  if (path === '/v2') return '/'
+  if (path.startsWith('/v2/')) return path.slice(3) || '/'
+  return path
+}
