+++ packages/phlo-observatory/src/phlo_observatory/vite.config.ts
@@ -32,6 +32,23 @@ const config = defineConfig({
   ssr: {
     noExternal: ['@primer/react'],
   },
+  build: {
+    rollupOptions: {
+      output: {
+        manualChunks(id) {
+          if (!id.includes('node_modules')) return undefined
+          if (id.includes('@primer/')) return 'vendor-primer'
+          if (id.includes('@xyflow/')) return 'vendor-flow'
+          if (id.includes('@base-ui/')) return 'vendor-base-ui'
+          if (id.includes('@tanstack/')) return 'vendor-tanstack'
+          if (id.includes('/react/') || id.includes('/react-dom/')) {
+            return 'vendor-react'
+          }
+          return undefined
+        },
+      },
+    },
+  },
   plugins: [
     devtools(),
     nitro({
